requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyJWT==2.8.0
cryptography==41.0.7
